Move CoreData funcs into DataManager+CoreData
